Move ProjectCard to common folder
